+++ docs/import/add_disaster_startup.js
@@ -1,5 +1,4 @@
 import {authenticateToFirebase, Authenticator} from '../authenticate.js';
-import {colorMap} from '../firebase_layers.js';
 import {loadNavbarWithTitle} from '../navbar.js';
 import SettablePromise from '../settable_promise.js';
 import TaskAccumulator from '../task_accumulator.js';
@@ -41,43 +40,4 @@ $('#tbody').sortable({
     return helper;
   },
 });
-
-function createRadioFor(colorType) {
-  const buttonAndLabel = [];
-  buttonAndLabel.push($(document.createElement('input')).attr({
-    name: 'color-type',
-    type: 'radio',
-    id: colorType + '-radio',
-    value: colorType,
-  }));
-  buttonAndLabel.push($(document.createElement('label'))
-                          .prop('for', colorType + 'radio')
-                          .text(colorType));
-  buttonAndLabel.push($(document.createElement('span')).text('  '));
-  return buttonAndLabel;
-}
-
-const colorFunctionDiv = $('#color-fxn-editor');
-colorFunctionDiv.prepend(createRadioFor('single-color'));
-colorFunctionDiv.prepend(createRadioFor('discrete'));
-colorFunctionDiv.prepend(createRadioFor('continuous'));
-
-function createColorPicker() {
-  const colorPicker = $(document.createElement('select'));
-  colorMap.forEach((value, key) => {
-    const option = $(document.createElement('option')).val(key).text(key);
-    colorPicker.append(option);
-  });
-  return colorPicker;
-}
-
-const label = $(document.createElement('label'))
-                  .prop('for', 'single-color-picker')
-                  .text('color: ');
-const singleColorPicker = createColorPicker().prop('id', 'single-color-picker');
-singleColorPicker.on('change', () => {
-  getRowIndex(singleColorPicker.parents('tr'))
-})
-$('#single').append(label, singleColorPicker);
-
 loadNavbarWithTitle('Add disaster');
